refactor to extract shared logic from logging.ts
